fix: addon position
Reactstrap removed their Addon component and changed their
implementation to position addons based on the order of elements. Our
Addon component still provides a position property, but it doesn't work.

Removed Addon and AddonButton components as the Addon became a fragment
wrapper and AddonButton is just a Button.
Added addonPosition property to Input component to position the addon
left or right.
